requirements.txt
pandas
snowflake-snowpark-python
streamlit
plotly